create base class for dashboard layout
